requirements.txt
PyQt5
pytesseract
Pillow
pygetwindow
pyautogui
pyperclip
